+++ src/vue/sheets/actor/character/InventoryTab.vue
@@ -5,7 +5,7 @@ import { ActorSheetContext, RootContext } from '@/vue/SheetContext';
 import CharacterDataModel from '@/actor/data/CharacterDataModel';
 import EquipmentDataModel, { EquipmentState } from '@/item/data/EquipmentDataModel';
 import GenesysItem from '@/item/GenesysItem';
-import { DragTransferData } from '@/data/DragTransferData';
+import { DragTransferData, extractDataFromDragTransferTypes } from '@/data/DragTransferData';
 import { transferInventoryBetweenActors } from '@/operations/TransferBetweenActors';
 
 import Localized from '@/vue/components/Localized.vue';
@@ -140,24 +140,15 @@ async function resetDragCounters(_event: DragEvent) {
 }
 
 function modifyDragCounters(event: DragEvent, direction: number) {
-	let draggedType: string;
-	const dragData = JSON.parse(event.dataTransfer?.getData('text/plain') ?? '{}') as DragTransferData;
-	if (dragData.genesysType) {
-		draggedType = dragData.genesysType;
-	} else if (dragData.uuid) {
-		const droppedEntity = fromUuidSync(dragData.uuid) as { type: string } | null;
-		if (!droppedEntity) {
-			return;
-		}
-		draggedType = droppedEntity.type;
-	} else {
+	const dragDataFromType = extractDataFromDragTransferTypes(event.dataTransfer?.types);
+	if (!dragDataFromType) {
 		return;
 	}
 
-	if (CharacterDataModel.isRelevantTypeForContext('EQUIPABLE', draggedType)) {
+	if (CharacterDataModel.isRelevantTypeForContext('EQUIPABLE', dragDataFromType.genesysType)) {
 		dragCounters.value[EquipmentState.Equipped] += direction;
 	}
-	if (CharacterDataModel.isRelevantTypeForContext('INVENTORY', draggedType)) {
+	if (CharacterDataModel.isRelevantTypeForContext('INVENTORY', dragDataFromType.genesysType)) {
 		dragCounters.value[EquipmentState.Carried] += direction;
 		dragCounters.value[EquipmentState.Dropped] += direction;
 	}